Medical - Hide cursor when unconscious (#8269)
* Hide cursor when blocking inputs

* Update addons/common/DisableMouseDialog.hpp


* Update addons/common/functions/fnc_disableUserInput.sqf


* Update addons/common/DisableMouseDialog.hpp
