+++ addons/common/config.cpp
@@ -14,18 +14,17 @@ class CfgPatches {
     };
 };
 
-#include "CfgEventHandlers.hpp"
-
-#include "CfgLocationTypes.hpp"
-#include "CfgSounds.hpp"
-#include "CfgVehicles.hpp"
-#include "CfgWeapons.hpp"
-#include "CfgMagazines.hpp"
-
-#include "CfgMoves.hpp"
-#include "CfgVoice.hpp"
-#include "CfgUnitInsignia.hpp"
 #include "CfgEden.hpp"
+#include "CfgEventHandlers.hpp"
+#include "CfgLocationTypes.hpp"
+#include "CfgMagazines.hpp"
+#include "CfgMoves.hpp"
+#include "CfgSounds.hpp"
+#include "CfgUnitInsignia.hpp"
+#include "CfgVehicles.hpp"
+#include "CfgVoice.hpp"
+#include "CfgWeapons.hpp"
+#include "CfgWrapperUI.hpp"
 
 class ACE_Rsc_Display_Base {
     idd = -1;
@@ -62,6 +61,7 @@ class ACE_Rsc_Control_Base {
 #include "ACE_Settings.hpp"
 #include "define.hpp"
 #include "ProgressScreen.hpp"
+#include "DisableMouseDialog.hpp"
 #include "HintConfig.hpp"
 #include "RscInfoType.hpp"
 #include "CompassControl.hpp"
